Remove the interactions data link.
This is solely to make the process of updating the results easier.

Currently:

1. Run results
2. Push
3. Tar the interactions
4. Manually upload to zenodo (which thankfully now has versioned DOI
   numbers).

While I really like having the data there with a DOI step 3 and 4
(especially 4) is just that little bit more of a pain and sadly can't be
automated (I believe).

If there are any objections I can keep doing this (it's not the highest
of workloads).